# server/src/test/resources/docs/ssapick-api-docs.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "SSAPICK API Documentation",
    "description" : "싸피 프로젝트 싸픽의 API 문서입니다.",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:8080/api/v1"
  } ],
...
  "paths" : {
...
    "/api/v1/follow/{userId}" : {
      "post" : {
        "tags" : [ "follow" ],
        "summary" : "로그인된 사용자가 다른 사용자를 팔로우한다.",
        "description" : "팔로우 생성 API",
        "operationId" : "follow-controller-test/팔로우_생성_성공_테스트",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "팔로우할 사용자 식별자",
          "required" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "201" : {
            "description" : "201",
            "content" : {
              "application/json;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-v1-follow-userId566855735"
                },
                "examples" : {
                  "follow-controller-test/팔로우_생성_성공_테스트" : {
                    "value" : "{\r\n  \"success\" : true,\r\n  \"status\" : 200,\r\n  \"message\" : \"success\",\r\n  \"data\" : null\r\n}"
                  }
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "follow" ],
        "summary" : "로그인된 사용자가 다른 사용자를 언팔로우한다.",
        "description" : "팔로우 삭제 API",
        "operationId" : "follow-controller-test/팔로우_삭제_성공_테스트",
        "parameters" : [ {
          "name" : "userId",
          "in" : "path",
          "description" : "언팔로우할 사용자 식별자",
          "required" : true,
          "schema" : {
            "type" : "number"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "204",
            "content" : {
              "application/json;charset=UTF-8" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-v1-follow-userId566855735"
                },
                "examples" : {
                  "follow-controller-test/팔로우_삭제_성공_테스트" : {
                    "value" : "{\r\n  \"success\" : true,\r\n  \"status\" : 200,\r\n  \"message\" : \"success\",\r\n  \"data\" : null\r\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "api-v1-follow-userId566855735" : {
        "type" : "object",
        "properties" : {
          "success" : {
            "type" : "boolean",
            "description" : "성공 여부"
          },
          "message" : {
            "type" : "string",
            "description" : "응답 메시지"
          },
          "status" : {
            "type" : "number",
            "description" : "응답 코드"
          }
        }
      },
...
    }
  }
}